fibonacci advanced

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,19 +8,20 @@
 
 int main(void)
 {
-	int prev = 1, curr = 2, count = 0;
+	long long prev = 1, curr = 2, count = 0;
 
-	printf("%d, %d", prev, curr);
+	printf("%lld, %lld", prev, curr);
 	count += 2;
 	while (count < 98)
 	{
-		int next = prev + curr;
+		long long next = prev + curr;
 
-		printf(", %d", next);
+		printf(", %lld", next);
 		count++;
 		prev = curr;
 		curr = next;
 	}
 	printf("\n");
+
 	return (0);
 }
